The core donut_decompress_block now does not auto advance address
It's less complex thus saves size and is faster.
It'll also allow easier manipulation of input addresses by higher layers, for example a block scheduler that saves an address of a block that'll be dupplicated later in the stream.

The downside is that the feature of just simply calling decompress_block twice for 128 bytes will not work anymore

donut_bulk_load_ayx is the new name of donut_block_ayx, and it works the same way as before, though adapted for the new semantics.